pyproject.toml
[project]
name = "zotero-mcp"
version = "0.1.0"
description = "MCP server for Zotero integration"
dependencies = [
    "fastmcp",
    "pyzotero>=1.5.0",
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "anthropic>=0.3.0",
    "pypdf>=3.0.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
